Stage_Test/MagicLinkProject/terraform/modules/cloud_function/function_files/gemini_complicated/requirements.txt
functions-framework==3.*
flask==2.*
requests==2.*
google-auth==2.*
google-cloud-firestore==2.*

